✅ feat: #59 PathResultからメッセージ出力
close #76
