+++ frontend/src/components/editor/panel-list/placement-controll-part.tsx
@@ -5,6 +5,8 @@ import { gridSlice } from "../../../store/slices/grid-slice";
 import { RootState } from "../../../store";
 import { useDispatch, useSelector } from "react-redux";
 
+import { PlaySimulateAsync } from "./play-simulate";
+
 export const PlacementControllPart: React.FC = () => {
   const dispatch = useDispatch();
   const gridHistory = useSelector((state: RootState) => state.grid.gridHistory);
@@ -33,22 +35,7 @@ export const PlacementControllPart: React.FC = () => {
 
   // 「再生」メソッド
   const playSimulation = async () => {
-    try {
-      console.log(grid);
-      // Gridを引数にしてAPIを呼び出す
-      const response = await fetch('/api/judge', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json',},
-        body: JSON.stringify(grid),
-      });
-      if (!response.ok) {
-        throw new Error(`HTTP error! Status: ${response.status}`);
-      }
-      const data = await response.json();
-      console.log('API疎通確認成功:', data);
-    } catch (error) {
-      console.error('API疎通確認エラー:', error);
-    }
+      await PlaySimulateAsync(grid);
   };
