+++ frontend/src/schemas/path.json
@@ -26,6 +26,8 @@
     },
     "Result": {
       "enum": [
+        "NoStart",
+        "NoGoal",
         "NoPath",
         "HasClearPath",
         "HasFailPath"
