+++ frontend/src/types/path/schema.ts
@@ -1,8 +1,18 @@
-export enum Result{
+export enum Result {
+  NoStart = "NoStart",
+  NoGoal = "NoGoal",
   NoPath = "NoPath",
   HasClearPath = "HasClearPath",
   HasFailPath = "HasFailPath",
 }
+
+export const resultMessages: Record<Result, string> = {
+  [Result.NoStart]: "スタート地点がありません",
+  [Result.NoGoal]: "ゴール地点がありません",
+  [Result.NoPath]: "経路がありませんでした",
+  [Result.HasClearPath]: "ゴールできました！🎉",
+  [Result.HasFailPath]: "間違った道でした...😢",
+};
 
 export type Vector = {
   x: number;
